+++ components/Layout/Navigation/AppBar.tsx
@@ -12,8 +12,8 @@ import {
 } from "@mui/material";
 import Image from "next/image";
 import { useRouter } from "next/router";
-import { useSession } from "../../lib/client/useSession";
-import { capitalizeFirstLetter } from "../ItemPopup";
+import { useSession } from "../../../lib/client/useSession";
+import { capitalizeFirstLetter } from "../../ItemPopup";
 import { UpdateButton } from "./UpdateButton";
 import InviteButton from "./UserMenu";
